[SofaKernel] Move SofaHelperTesting into SofaFramework
