+++ SofaKernel/SofaFramework/SofaHelperTesting/src/sofa/helper/testing/BaseTest.h
@@ -27,7 +27,7 @@
 #ifndef SOFA_BASETEST_H
 #define SOFA_BASETEST_H
 
-#include "SofaGtest.h"
+#include <sofa/helper/testing/SofaGtest.h>
 #include <sofa/helper/testing/TestMessageHandler.h>
 
 namespace sofa {
