+++ SofaKernel/SofaFramework/SofaHelperTesting/src/sofa/helper/testing/SofaGtest.h
