--- SofaKernel/SofaHelperTesting/CMakeLists.txt
@@ -1,27 +1,0 @@
-cmake_minimum_required(VERSION 3.1)
-project(SofaHelperTesting)
-
-find_package(SofaFramework REQUIRED)
-find_package(gtest REQUIRED)
-
-set(HEADER_FILES
-    src/sofa/helper/testing/SofaGtest.h
-    src/sofa/helper/testing/BaseTest.h
-    src/sofa/helper/testing/NumericTest.h
-    src/sofa/helper/testing/TestMessageHandler.h
-)
-
-set(SOURCE_FILES
-    src/sofa/helper/testing/BaseTest.cpp
-    src/sofa/helper/testing/NumericTest.cpp
-    src/sofa/helper/testing/TestMessageHandler.cpp
-)
-
-add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
-
-target_link_libraries(SofaHelperTesting SofaHelper gtest)
-target_include_directories(SofaHelperTesting PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src ${SofaFramework_INCLUDE_DIRS} ${gtest_INCLUDE_DIRS})
-
-sofa_create_package(SofaHelperTesting 1.0 SofaHelperTesting src)
-
-
